Portfolio Section

+++ src/controllers/clientController.js
@@ -1,46 +1,16 @@
-import Client from "../models/Clients";
-import models from "../models";
+import Client from "../models/Client";
+import Portfolio from "../models/Portfolio";
+
 export default {
-  /*     list: async (req, res, next) => {
-        try {
-            const result = await Client.find()
-            .populate('Services', {name: 1});
-            res.status(200).json(result);
-        }
-        catch (e) {
-            res.status(500).send({
-                message: 'Ocurrió un error'
-            });
-            next(e);
-        }
-    }, */
-  /*   list: async (req, res, next) => {
-    try {
-      let value = req.query.value;
-      const reg = await models.Clients.find().populate("services.service", {
-        name: 1,
-        _id: 0,
-        price: 1,
-      }); //tiene que ir en minúsuclas, no preguntes porque.. jeje
-      //.populate('services').populate('service', { name: 1, _id: 0, price: 1 });
-      res.status(200).json(reg);
-    } catch (e) {
-      res.status(500).send({
-        message: "Ocurrió un error",
-      });
-      next(e);
-    }
-  }, */
   list: async (req, res, next) => {
     try {
-      //let value = req.query.value;
-      const reg = await models.Clients.find().populate("services", {
+      const reg = await Client.find().populate("services", {
         name: 1,
         _id: 1,
         price: 1,
+        serviceType: 1,
         description: 1,
-      }); //tiene que ir en minúsuclas, no preguntes porque.. jeje
-      //.populate('services').populate('service', { name: 1, _id: 0, price: 1 });
+      });
       res.status(200).json(reg);
     } catch (e) {
       res.status(500).send({
@@ -51,19 +21,17 @@ export default {
   },
   add: async (req, res, next) => {
     try {
-      const { name, lastname, email, phone, services, address } = req.body;
+      const { name, lastname, email, phone, services, company } = req.body;
       const newClient = new Client({
         name,
         lastname,
         email,
         phone,
         services,
-        address,
+        company,
       });
       const clientSaved = await newClient.save();
       res.status(200).json(clientSaved);
-      //const reg = await models.Post.create(req.body);
-      //res.status(200).json(reg);
     } catch (error) {
       console.log(error);
       res.status(500).send({
@@ -78,7 +46,7 @@ export default {
         { _id: req.body._id },
         req.body,
         { new: true }
-      ); //el new es para que me devuelva los datos actualizados, no el registro viejo.
+      );
       res.status(200).json(clientUpdated);
     } catch (error) {
       res.status(500).send({
@@ -149,12 +117,49 @@ export default {
   },
   deleteClientById: async (req, res, next) => {
     try {
-      const reg = await Client.findByIdAndDelete({ _id: req.query.id });
-      res.status(200).json(reg);
+      const Portfolios = await Portfolio.find({
+        client: req.query.id,
+      });
+
+      if (Portfolios.length == 0) {
+        const reg = await Client.findByIdAndDelete({ _id: req.query.id });
+        res.status(200).json(reg);
+      }
+
+      res.status(500).json({
+        message:
+          "El cliente tiene un Portfolio activo, elimine primero el portoflio.",
+      });
     } catch (error) {
       res.status(500).send({
         message: "Ocurrió un error.",
       });
+      next(error);
+    }
+  },
+  getMonthlyPayments: async (req, res, next) => {
+    try {
+      const result = await Client.find().populate({
+        path: "services",
+        match: { serviceType: "Mensual" },
+      });
+      res.status(200).json(result);
+    } catch (error) {
+      console.log(error);
+      res.status(500).send({ message: "Ocurrió un error" });
+      next(error);
+    }
+  },
+  getAnnualPayments: async (req, res, next) => {
+    try {
+      const result = await Client.find().populate({
+        path: "services",
+        match: { serviceType: "Anual" },
+      });
+      res.status(200).json(result);
+    } catch (error) {
+      console.log(error);
+      res.status(500).send({ message: "Ocurrió un error" });
       next(error);
     }
   },
